feat: move restriction to metadmodel
